pyproject.toml
[project]
name = "polymoly"
version = "0.1.0"
description = "Advanced Polymarket High-Frequency Algorithmic Trading System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.2.0",
    "pymc>=5.10.0",
    "econml>=0.14.0",
    "scikit-learn>=1.3.0",
    "pandas>=2.1.0",
    "websockets>=12.0",
    "requests>=2.31.0",
    "numpy>=1.26.0",
    "py-clob-client>=0.1.0",
    "tabulate>=0.9.0",
    "tqdm>=4.66.0",
    "scipy>=1.11.0",
    "python-dotenv>=1.0.0"
]

[tool.setuptools.packages.find]
include = ["algos*"]
